Redesign Tree view to support lazy loading
