+++ src/pages/ExperimentTree.jsx
@@ -1,7 +1,8 @@
 import { useEffect, useMemo, useRef, useState, MutableRefObject } from "react";
 import { autosubmitApiV3 } from "../services/autosubmitApiV3";
-import { useParams } from "react-router-dom";
-import FancyTree from "../common/FancyTree";
+import { autosubmitApiV4 } from "../services/autosubmitApiV4";
+
+import { Link, useParams } from "react-router-dom";
 import useASTitle from "../hooks/useASTitle";
 import useBreadcrumb from "../hooks/useBreadcrumb";
 import JobDetailCard from "../common/JobDetailCard";
@@ -11,9 +12,149 @@ import TreeContentHandler from "../components/context/tree/business/treeUpdate";
 import BottomPanel from "../common/BottomPanel";
 import { ChangeStatusModal } from "../common/ChangeStatusModal";
 
+/***
+ * param treeData: Array of objects representing the tree structure
+ */
+const getTreeDirs = (treeData) => {
+  const result = [];
+
+  // First level: date
+  Object.keys(treeData).forEach((date) => {
+    const dateNode = {
+      title: date,
+      key: date,
+      date: date,
+      member: "NA",
+      section: null,
+      chunk: "NA",
+      children: [],
+    };
+
+    let date_statuses_counters = {};
+
+    // Second level: member
+    Object.keys(treeData[date]).forEach((member) => {
+      const memberNode = {
+        title: member,
+        key: `${date}-${member}`,
+        date: date,
+        member: member,
+        section: null,
+        chunk: "NA",
+        children: [],
+      };
+
+      let member_statuses_counters = {};
+
+      // Third level: section
+      Object.keys(treeData[date][member]).forEach((section) => {
+        let section_statuses_counters = treeData[date][member][section];
+
+        let jobs_counter = 0;
+
+        Object.entries(section_statuses_counters).forEach(([status, count]) => {
+          jobs_counter += count;
+          if (member_statuses_counters[status]) {
+            member_statuses_counters[status] += count;
+          } else {
+            member_statuses_counters[status] = count;
+          }
+        });
+
+        // Check if the section has more than one job to determine if it's a directory
+        // If it has more than one job, we treat it as a directory; otherwise,
+        // we treat it as a leaf node that won't be displayed in the tree dir view.
+        const is_section_dir = jobs_counter > 1;
+        if (is_section_dir) {
+          const sectionNode = {
+            title: section,
+            key: `${date}-${member}-${section}`,
+            date: date,
+            member: member,
+            section: section,
+            chunk: null,
+            children: [],
+            statuses_counters: section_statuses_counters,
+          };
+
+          memberNode.children.push(sectionNode);
+        }
+      });
+
+      memberNode.statuses_counters = member_statuses_counters;
+
+      // Update date level statuses counters
+      Object.entries(member_statuses_counters).forEach(([status, count]) => {
+        if (date_statuses_counters[status]) {
+          date_statuses_counters[status] += count;
+        } else {
+          date_statuses_counters[status] = count;
+        }
+      });
+
+      if (member === "NA") {
+        // Forward the children of the memberNode to the dateNode if member is "NA"
+        dateNode.children.push(...memberNode.children);
+      } else {
+        dateNode.children.push(memberNode);
+      }
+    });
+
+    dateNode.statuses_counters = date_statuses_counters;
+    if (date === "NA") dateNode.title = "Keys";
+
+    result.push(dateNode);
+  });
+
+  console.log(JSON.stringify(result, null, 2));
+  return result;
+};
+
+const HierarchyTree = ({ treeDirs, onNodeSelect }) => {
+  const handleNodeSelect = (node) => {
+    if (onNodeSelect) {
+      onNodeSelect(node);
+    }
+  };
+
+  return (
+    <ul>
+      {treeDirs &&
+        treeDirs.length > 0 &&
+        treeDirs.map((treeDir, index) => (
+          <li key={index}>
+            <span
+              className="cursor-pointer hover:bg-black/10 "
+              onClick={() => handleNodeSelect(treeDir)}
+            >
+              {treeDir.title}{" "}
+              <span style={{ fontSize: "0.8em", color: "gray" }}>
+                {treeDir.statuses_counters &&
+                  Object.entries(treeDir.statuses_counters).map(
+                    ([status, count]) => (
+                      <span key={status} style={{ marginRight: "5px" }}>
+                        {status}: {count}
+                      </span>
+                    ),
+                  )}
+              </span>
+            </span>
+            {treeDir.children && treeDir.children.length > 0 && (
+              <div style={{ marginLeft: "20px" }}>
+                <HierarchyTree
+                  treeDirs={treeDir.children}
+                  onNodeSelect={handleNodeSelect}
+                />
+              </div>
+            )}
+          </li>
+        ))}
+    </ul>
+  );
+};
+
 const ExperimentTree = () => {
   const dispatch = useDispatch();
-  const authState = useSelector((state) => state.auth);
   const routeParams = useParams();
   useASTitle(`Experiment ${routeParams.expid} tree`);
   useBreadcrumb([
@@ -27,326 +168,83 @@ const ExperimentTree = () => {
     },
   ]);
 
-  /** @type {MutableRefObject<Fancytree.Fancytree>} */
-  const tree = useRef();
-  const filterRef = useRef();
+  const [selectedDir, setSelectedDir] = useState(null);
 
-  const [activeMonitor, setActiveMonitor] = useState(false);
-
-  const [jobs, setJobs] = useState([]);
-  const [selectedJobIds, setSelectedJobIds] = useState([]);
-  const selectedJob = useMemo(() => {
-    if (selectedJobIds.length === 1) {
-      return jobs.find((rawNode) => rawNode.label === selectedJobIds[0]);
-    }
-    return undefined;
-  }, [jobs, selectedJobIds]);
-
-  const [showRunsM, setShowRunsM] = useState(false);
-  const [selectedRun, setSelectedRun] = useState({
-    run_id: null,
-    created: null,
-  });
-
-  const [showModal, setShowModal] = useState(false);
-  const toggleModal = ( refresh = false ) => {
-    setShowModal(!showModal);
-    if (refresh === true) {
-      refetch();
-    }
-  };
-
-  const { data, isFetching, refetch, isError } =
-    autosubmitApiV3.endpoints.getExperimentTreeView.useQuery({
+  const { data: jobsCategoryTreeData } =
+    autosubmitApiV4.endpoints.getJobsCategoryTree.useQuery({
       expid: routeParams.expid,
-      runId: selectedRun?.run_id,
     });
 
-  const { data: pklData } = autosubmitApiV3.endpoints.getPklTreeInfo.useQuery(
-    {
-      expid: routeParams.expid,
-    },
-    {
-      skip: !activeMonitor || selectedRun?.run_id,
-      pollingInterval: 5 * 1000,
-    }
-  );
+  const tree = useMemo(() => {
+    if (jobsCategoryTreeData && jobsCategoryTreeData.category_tree)
+      return getTreeDirs(jobsCategoryTreeData.category_tree);
 
-  const assignKeysToTree = (tree) => {
-    if (Array.isArray(tree)) {
-      // Create a deep copy and assign keys
-      return tree.map((node) => {
-        const newNode = { ...node };
-        newNode.key = `${node.refKey}-${Math.random().toString(36).slice(2, 11)}`;
-        if (Array.isArray(newNode.children)) {
-          newNode.children = assignKeysToTree(newNode.children);
-        }
-        return newNode;
-      });
-    }
+    return [];
+  }, [jobsCategoryTreeData]);
+
+  const handleDirSelect = (dir) => {
+    setSelectedDir(dir);
   };
 
-  const dataTree = useMemo(() => assignKeysToTree(data?.tree || []), [data]);
-
-  useEffect(() => {
-    // Unmount component
-    return () => {
-      const promise = dispatch(
-        autosubmitApiV3.endpoints.showdownRoute.initiate(
-          {
-            route: "tree",
-            loggedUser: authState.user_id,
-            expid: routeParams.expid,
-          },
-          { forceRefetch: true }
-        )
-      );
-      promise.unsubscribe();
-    };
-    // eslint-disable-next-line
-  }, []);
-
-  useEffect(() => {
-    const newJobs = data?.jobs || [];
-    setJobs(newJobs);
-  }, [data]);
-
-  useEffect(() => {
-    let treeUpdater = new TreeContentHandler(
-      jobs,
-      data?.reference || {},
-      pklData,
-      false
+  const { data: jobsData, isFetching: isJobsFetching } =
+    autosubmitApiV4.endpoints.getExperimentJobs.useQuery(
+      {
+        expid: routeParams.expid,
+        date: selectedDir?.date || undefined,
+        member: selectedDir?.member || undefined,
+        section: selectedDir?.section || undefined,
+        chunk: selectedDir?.chunk || undefined,
+      },
+      {
+        skip: !selectedDir, // Skip fetching if no directory is selected
+      },
     );
-
-    if (tree.current && treeUpdater.validate()) {
-      const processResult = treeUpdater.processChanges(tree.current);
-      if (Array.isArray(processResult?.currentJobs)) {
-        setJobs(processResult.currentJobs);
-      }
-    }
-  }, [pklData]);
-
-  const toggleActiveMonitor = () => {
-    setActiveMonitor(!activeMonitor);
-  };
-
-  const handleFilter = (e) => {
-    e.preventDefault();
-    if (tree.current) {
-      if (filterRef.current.value) {
-        tree.current.filterNodes(filterRef.current.value);
-      } else {
-        tree.current.clearFilter();
-      }
-    }
-  };
-
-  const handleClear = () => {
-    filterRef.current.value = "";
-    if (tree.current) tree.current.clearFilter();
-  };
-
-  const handleSelectNodes = (_selectedNodes) => {
-    if (Array.isArray(_selectedNodes)) {
-      const newSelectedIds = _selectedNodes
-        .filter((node) => node?.folder === undefined)
-        .map((node) => node?.refKey);
-      setSelectedJobIds(newSelectedIds);
-    } else {
-      setSelectedJobIds([]);
-    }
-  };
-
-  const handleExpand = () => {
-    if (tree.current) tree.current.expandAll();
-  };
-  const handleCollapse = () => {
-    if (tree.current) tree.current.expandAll(false);
-  };
-  const handleDefaultExpand = () => {
-    if (tree.current) {
-      tree.current.expandAll(false, {
-        noAnimation: true,
-        noEvents: true,
-      });
-      tree.current.getRootNode().children.forEach((node) => {
-        node.setExpanded();
-      });
-    }
-  };
-
-  const handleRunSelect = (item) => {
-    handleClear();
-    setSelectedRun(item);
-    setShowRunsM(false);
-  };
-
-  const handleResetRun = () => {
-    handleRunSelect({
-      run_id: null,
-      created: null,
-    });
-  };
-
-  const toggleShowRunsM = () => {
-    setShowRunsM(!showRunsM);
-  };
 
   return (
     <>
-      <RunsModal
-        show={showRunsM}
-        onHide={toggleShowRunsM}
-        expid={routeParams.expid}
-        onRunSelect={handleRunSelect}
-      />
       <div className="w-full flex flex-col gap-4 grow">
-        {(isError || data?.error) && (
-          <span className="alert alert-danger rounded-2xl">
-            <i className="fa-solid fa-triangle-exclamation me-2"></i>{" "}
-            {data?.error_message || "Unknown error"}
-          </span>
-        )}
+        <Link to={`/experiment/${routeParams.expid}/tree-legacy`}>
+          Go to Legacy Tree View
+        </Link>
 
-        <div className="flex gap-2 items-center flex-wrap">
-          <div>
-            <button
-              className={
-                "btn btn-primary font-bold " +
-                (selectedRun?.run_id && " rounded-e-none")
-              }
-              title="Select Run"
-              onClick={toggleShowRunsM}
-            >
-              <i className="fa-solid fa-clock-rotate-left me-2"></i> Run:{" "}
-              {(selectedRun.run_id && selectedRun.created) || "Latest"}
-            </button>
-            {selectedRun?.run_id && (
-              <button
-                className="btn btn-light text-primary rounded-s-none border"
-                title="Reset to latest"
-                onClick={handleResetRun}
-              >
-                <i className="fa-solid fa-xmark"></i>
-              </button>
+        <div className="flex gap-4 grow">
+          <div className="flex flex-col border border-gray-300 rounded-lg p-4 w-1/4">
+            <HierarchyTree treeDirs={tree} onNodeSelect={handleDirSelect} />
+          </div>
+
+          <div className="flex flex-col border border-gray-300 rounded-lg p-4 w-3/4">
+            {selectedDir ? (
+              <div>
+                <h3>Selected Directory Details</h3>
+                <p>Date: {selectedDir.date || "Any"}</p>
+                <p>Member: {selectedDir.member || "Any"}</p>
+                <p>Section: {selectedDir.section || "Any"}</p>
+              </div>
+            ) : (
+              <p>Please select a directory from the tree to see its details.</p>
             )}
-          </div>
-          <form className="grow flex flex-wrap" onSubmit={handleFilter}>
-            <input
-              ref={filterRef}
-              className="form-input rounded-e-none grow"
-              placeholder="Filter job..."
-            />
-            <button
-              type="submit"
-              className="btn btn-dark border-dark font-bold px-4 rounded-s-none rounded-e-none"
-            >
-              Filter
-            </button>
-            <button
-              type="button"
-              className="btn btn-light border font-bold px-4 rounded-s-none"
-              onClick={handleClear}
-            >
-              Clear
-            </button>
-          </form>
-          {!selectedRun?.run_id && (
-            <button
-              className={
-                "btn font-bold px-4 text-nowrap " +
-                (activeMonitor ? "btn-danger" : "btn-success")
-              }
-              onClick={toggleActiveMonitor}
-            >
-              {activeMonitor ? "STOP MONITORING" : "START MONITOR"}
-            </button>
-          )}
-          <button
-            className="btn btn-success font-bold"
-            title="Refresh data"
-            onClick={() => {
-              refetch();
-            }}
-          >
-            <i className="fa-solid fa-rotate-right"></i>
-          </button>
-        </div>
 
-        <div className="flex flex-wrap gap-2 items-center justify-between">
-          <span className="mx-2 text-sm">
-            Total #Jobs: {data?.total} | Chunk unit:{" "}
-            {data?.reference?.chunk_unit} | Chunk size:{" "}
-            {data?.reference?.chunk_size}
-          </span>
-          <div className="flex gap-2">
-            <button
-              className="btn btn-success font-bold px-4 text-sm"
-              onClick={handleDefaultExpand}
-            >
-              Default Expand
-            </button>
-            <button
-              className="btn btn-primary font-bold px-4 text-sm"
-              onClick={handleExpand}
-            >
-              Expand All +
-            </button>
-            <button
-              className="btn btn-secondary font-bold px-4 text-sm"
-              onClick={handleCollapse}
-            >
-              Collapse All -
-            </button>
+            {
+              jobsData && jobsData.jobs && jobsData.jobs.length > 0 ? (
+                <div className="mt-4">
+                  <h4>Jobs in Selected Directory</h4>
+                  <ul>
+                    {jobsData.jobs.map((job) => (
+                      <li key={job.job_name}>
+                        {job.name} - Status: {job.status}
+                      </li>
+                    ))}
+                  </ul>
+                </div>
+              ) : (
+                selectedDir && (
+                  <p>No jobs found for the selected directory.</p>
+                )
+              )
+            }
           </div>
-        </div>
-
-        <div className="relative grow basis-0 overflow-auto min-h-[70vh] lg:min-h-[50vh] w-full border p-4 rounded-lg custom-scrollbar bg-white">
-          {isFetching && (
-            <div className="absolute inset-0 z-20 flex items-center justify-center bg-white">
-              <div className="spinner-border dark:invert" role="status"></div>
-            </div>
-          )}
-          <FancyTree
-            tree={(_tree) => {
-              tree.current = _tree;
-            }}
-            source={dataTree}
-            onSelectNodes={handleSelectNodes}
-          />
         </div>
       </div>
-
-      {selectedJobIds.length > 0 && (
-        <BottomPanel
-          title={
-            selectedJobIds.length === 1
-              ? selectedJobIds[0]
-              : `${selectedJobIds.length} jobs selected`
-          }
-        >
-          <div className="flex flex-col gap-3">
-            <JobDetailCard
-              expid={routeParams.expid}
-              jobData={selectedJob}
-              jobs={jobs}
-            />
-            <div className="flex items-center justify-center gap-3">
-              <div className="font-semibold">Actions:</div>
-              <button className="btn btn-primary" onClick={toggleModal}>
-                Change status
-              </button>
-            </div>
-            <ChangeStatusModal
-              selectedJobs={selectedJobIds}
-              show={showModal}
-              onHide={toggleModal}
-              expid={routeParams.expid}
-            />
-          </div>
-        </BottomPanel>
-      )}
     </>
   );
 };
