+++ src/services/autosubmitApiV4.js
@@ -46,10 +46,16 @@ export const autosubmitApiV4 = createApi({
             }
         }),
         getExperimentJobs: builder.query({
-            query: ({ expid }) => {
+            query: ({ expid, date = undefined, member = undefined, section = undefined, chunk = undefined }) => {
                 return {
                     url: `experiments/${expid}/jobs`,
-                    method: "GET"
+                    method: "GET",
+                    params: {
+                        date,
+                        member,
+                        section,
+                        chunk
+                    }
                 }
             }
         }),
@@ -242,6 +248,12 @@ export const autosubmitApiV4 = createApi({
                 params: {
                     include_status: include_status
                 }
+            }),
+        }),
+        getJobsCategoryTree: builder.query({
+            query: ({ expid }) => ({
+                url: `experiments/${expid}/jobs-category-tree`,
+                method: "GET"
             }),
         }),
     }),
